Add "zyro.ga" example files via upload
These files are the backup archives for our example Zyro site, zyro.ga. These were created on a premium account before Zyro was phased out on October 1, 2020, and are as basic as it gets (based on a Zyro template). The .sitebuilder file can only be opened with WinRar on Windows (that we know of), and contains raw JSON that we're looking to parse. The zip file contains all the files inside the .sitebuilder file, for those that can't open it.